requirements.txt
anthropic>=0.7.0
slack_sdk>=3.0.0
imageio>=2.30.0
opencv-python-headless
scikit-image
numpy
python-dotenv
requests
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
reportlab>=4.0.0

# V2 CAPTCHA Multi-Provider Support
openai>=1.0.0              # GPT-4o CAPTCHA solving + Whisper audio transcription (recommended primary)
google-generativeai>=0.3.0  # Gemini CAPTCHA solving (optional alternative)